SUBUI-3: Split Up HttpClient Classes
