+++ src/Http/Headers/IHttpClientHeader.ts
@@ -1,0 +1,5 @@
+export default interface IHttpClientHeader {
+    readonly name: string;
+    readonly value: string;
+    toRecord(): Record<string, string>
+}
