+++ src/Http/HttpClientMethod.ts
@@ -1,0 +1,8 @@
+class HttpClientMethod {
+    public static GET: string = "GET";
+    public static POST: string = "POST";
+    public static PUT: string = "PUT";
+    public static DELETE: string = "DELETE";
+}
+
+export default HttpClientMethod;
