+++ src/Http/HttpClientReponseStatusCode.ts
@@ -1,0 +1,12 @@
+enum HttpClientResponseStatusCode {
+    OK = 200,
+    Created = 201,
+    NoContent = 204,
+    BadRequest = 400,
+    Unauthorized = 401,
+    Forbidden = 403,
+    NotFound = 404,
+    Conflict = 409
+}
+
+export default HttpClientResponseStatusCode;
